+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ## [1.0.0]
 * context
-* use `React.Children.map` to inject selected property
+* use `React.Children.map` to inject `selected` property
 
 ## [2.0.0]
 * uidList is given via props
+* inject `selected` via props by developer